Revert "membuat history"
