+++ nbproject/project.properties
@@ -28,13 +28,11 @@ dist.jar=${dist.dir}/TUBES_V5C.jar
 dist.javadoc.dir=${dist.dir}/javadoc
 excludes=
 file.reference.jcalendar-1.4.jar=../jcalendar-1.4.jar/jcalendar-1.4.jar
-file.reference.jcalendar-1.4.jar-1=D:\\Tugas Semester 3\\PBO\\Package_Jcalender\\lib\\jcalendar-1.4.jar
 includes=**
 jar.compress=false
 javac.classpath=\
     ${libs.MySQLDriver.classpath}:\
-    ${file.reference.jcalendar-1.4.jar}:\
-    ${file.reference.jcalendar-1.4.jar-1}
+    ${file.reference.jcalendar-1.4.jar}
 # Space-separated list of extra javac options
 javac.compilerargs=
 javac.deprecation=false
